requirements.txt
twitchio>=2.6.0
asqlite>=0.19.0
python-dotenv>=1.0.0 